# app/ui/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "filmpaw-server",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/sources": {
   "get": {
    "summary": "List Sources",
    "operationId": "list_sources_api_sources_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/SourceOut"
         },
         "type": "array",
         "title": "Response List Sources Api Sources Get"
        }
       }
      }
     }
    }
   },
   "post": {
    "summary": "Add Source",
    "operationId": "add_source_api_sources_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SourceIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SourceCreated"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/sources/{source_id}": {
   "delete": {
    "summary": "Delete Source",
    "operationId": "delete_source_api_sources__source_id__delete",
    "parameters": [
     {
      "name": "source_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Source Id"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/sources/{source_id}/scan": {
   "post": {
    "summary": "Scan One",
    "operationId": "scan_one_api_sources__source_id__scan_post",
    "parameters": [
     {
      "name": "source_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Source Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ScanSummary"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "ScanSummary": {
    "properties": {
     "added": {
      "type": "integer",
      "title": "Added"
     },
     "refreshed": {
      "type": "integer",
      "title": "Refreshed"
     },
     "missing": {
      "type": "integer",
      "title": "Missing"
     }
    },
    "type": "object",
    "required": [
     "added",
     "refreshed",
     "missing"
    ],
    "title": "ScanSummary"
   },
...
   "SourceCreated": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "unc_path": {
      "type": "string",
      "title": "Unc Path"
     },
     "label": {
      "type": "string",
      "title": "Label"
     }
    },
    "type": "object",
    "required": [
     "id",
     "unc_path",
     "label"
    ],
    "title": "SourceCreated"
   },
   "SourceIn": {
    "properties": {
     "unc_path": {
      "type": "string",
      "title": "Unc Path"
     },
     "label": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Label"
     }
    },
    "type": "object",
    "required": [
     "unc_path"
    ],
    "title": "SourceIn"
   },
   "SourceOut": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "unc_path": {
      "type": "string",
      "title": "Unc Path"
     },
     "label": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Label"
     },
     "last_scan_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Scan At"
     },
     "performer_count": {
      "type": "integer",
      "title": "Performer Count"
     },
     "reachable": {
      "type": "boolean",
      "title": "Reachable"
     }
    },
    "type": "object",
    "required": [
     "id",
     "unc_path",
     "label",
     "last_scan_at",
     "performer_count",
     "reachable"
    ],
    "title": "SourceOut"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  }
 }
}